Move functions to OfflinePersistence, macro out some debug functions, cleanup
